refactor : 기존의 RedisTemplate를 사용해 인증코드를 저장-조회하는 부분을 VerificationCodeRepository로 대체 (#45)
